Guard against static members overlapping due to traits
Summary: HHVM will fatal if you attempt to redeclare an instance member as static (or vice versa). We guard against this when you redeclare within a class, but *not* when you implicitly redeclare due to trait inclusion. This is a catch-all soln that looks for the intersection between static and instance methods / properties.

Reviewed By: CatherineGasnier

Differential Revision: D41360503

fbshipit-source-id: 0e4eafde11d30e48fd7175461578a9dccda5d9ca