Add support for a suppression file path
Summary: Our older autoloader checks the existence of a suppression file in the repo. If that file exists at the start of a request, we don't update or sync Facts for that request. This can be used if it would already be a bug for files to change in a given repo.

Reviewed By: DavidSnider

Differential Revision: D41822734

fbshipit-source-id: 54854a2c6ff83850209db9e30c00512186e9b5b7