Wrap socket pointer in ReleaseDeleter consistently between FizzAsyncServer and AsyncSSLSocket
Summary:
`apache::thrift::transport::detail::convertToShared` wraps the pointer with a `ReleaseDeleter` that seems to be expected in `Cpp2Connection`, when upgrading from Header to Rocket transport.

We weren't applying it for TLS 1.3/FIzz connections, leading to crashes in `TransportUpgradeSendCallback::messageSent` when trying to extract a `ReleaseDeleter` with `std::get_deleter`.

Reviewed By: praihan, mingtaoy

Differential Revision: D41814639

fbshipit-source-id: 0b2bd9d466beea0dd37636e53d3aff88374ffcc2