Fix use of WidthAnalysis in vasm-jumps
Summary: The only place we use vasm width information (other than Flags) for correctness is during `diamondIntoCmov` in vasm-jumps. After we've performed architecture specific lowering this information is no longer dependable (e.g. lower-x64 replaces `movtqb` instructions with `copy`s). Rather than disable the optimization we can simply emit `cmovq` when we can't be sure of the width, while this may be sub-optimal it should always be correct (and as we've already broken width invariants it should be safe).

Reviewed By: ricklavoie

Differential Revision: D41828468

fbshipit-source-id: 74d53e8d59f9f2a28af84e8042686242cd9a7bb4