remote old decl service: only fetch during init
Summary:
This is fine because when editing in the IDE, opening a file trigger decling of it so that old decls are made available. Therefore it's ok to use the old decl service only during init.

This is done simply by wiring down a `during_init` flag parameter to the shallow decl provider.

The changes in this diff are behind a flag, so we can make sure this does not impact fanout sizes.

Reviewed By: bobrenjc93

Differential Revision: D41841676

fbshipit-source-id: bf273910e6afa7beceff62a26ed6eb46499ae253