Support default values in lists of adapted types
Summary:
The cast in cxx_value_prefix is applied on the wrong side of Adapter::fromThrift in some cases. The cast was added to avoid type ambiguity for numeric types (D30782083), so move it to the location of the literals to ensure it is generated in the right place.

Ultimately the correct design here will separate treatment of type and field adapters instead of consolidating them via field:cpp_first_adapter. This diff moves us in that direction while avoiding the large full refactor.

Reviewed By: thedavekwon

Differential Revision: D41820314

fbshipit-source-id: 1f20a756f224770eea5864c990ce24379b96a688