Add a GitHub Action to build Ubuntu 20.04
This PR ports the scripts in
github.com/hhvm/packaging over to
GitHub Actions so we can build the HHVM
package from GitHub.

We run the build script in a container as
the "build-depends` Debian package triggers
some conflict with the packages included
in the default GitHub image.

The only modification to the packaging
scripts are in `make-debianish-package`:
1. The `$IS_NIGHTLY` checks were made consistent
as it caused compilation errors.
2. We now fetch the full repository instead of
using a source archive.
3. [minor] Simplified the `BUILD_DEPS_{OLD|NEW}`
that was unnecessary.
4. [minor] Several variables are derived from the
enviroment (`$DISTRO_ROOT`, `$OUT`, ...) instead
of being hardcoded.

This PR doesn't compile as the default GitHub runners
only have 7 Gb of memory and thus run out of memory:
https://docs.github.com/en/actions/using-github-hosted-runners/about-github-hosted-runners#supported-runners-and-hardware-resources
But I have tested the steps in isolation so it should
work with larger runners.