Add structured annotations
Summary:
Includes structured annotations for all definitions.
Adds the `StructuredAnnotation` struct to schema.thrift as the schema for these values, though this type is not used anywhere because the annotations are included in the generic value list.
wrap_with_protocol_value helper is in schematizer.h to avoid circular dependency issue with const_util.h

Reviewed By: vitaut

Differential Revision: D41597181

fbshipit-source-id: cbf13bab1b8ccf47d1bc833c86caab9d6c9ec5fd