hierarchy support
Summary:
Teach textual about the class hierarchy.

Currently we treat the base class, interfaces, and traits all as bases.  In addition we have the static versions of those as bases on the static class as well.

Added metadata about the class which is currently just emitted as comments in the type definition.

Reviewed By: davidpichardie

Differential Revision: D41676627

fbshipit-source-id: d429a23c3e9cf968c25afd24c80e5fa0a101f4c1