Properly Utilize Cache for Modules
Summary:
# Context
When modules were first added to the type checker in D34481050 (https://github.com/facebook/hhvm/commit/85b55e9c85c84c6cfb7c3937668912ce7197cd1e), a small bug was introduced in its Decl_provider. It did not check if the module was already contained in the Decl_store and instead would always fetch the file containing the module definition and parse it.

# Mitigation
To fix this I rewrote it in the same style as the other decl providers to check in the store before falling back to parsing.

Differential Revision: D41980289

fbshipit-source-id: 3e3642b764d2d8c4976ff63a6c652af2158ee82a