use the new hh_fanout typing_deps mode
Summary:
Plumbing to have the workers use the new `hh_fanout` to commit discovered edges.

We use the same strategy as the other `Typing_deps_mode`s: we keep a hash set of 1000 dependency edges in memory before committing and flushing them to disk via `HhFanout::commit_edges.`

This diff shouldn't impact legacy `hh_server` at all. The only public methods of `typing_deps.ml` that are affected are `add_idep` and `hh_fanout_flush_edges`. For `add_idep`, only the behavior when the mode is `HhFanoutRustMode` is changed in this diff.

`hh_worker` now takes the additional argument `--fanout-state-dir`, which behaves similarly as an identifier as `--decl-state-dir`.

Reviewed By: ljw1004

Differential Revision: D40420796

fbshipit-source-id: 407da495e8918ea97b580fb36be3d05bee0aa726