Treat generic parameter as bounded by mixed for type structure access
Summary: Accessing certain fields (e.g. `classname`) off a type structure should be permitted only if the type structure corresponds to a type for which that field is relevant. It's not possible to access such fields off a type structure derived from `mixed`. Nor is it possible to access it off a type structure derived from a generic whose bound is explicitly `mixed`. In this diff we fix things so that it's also not possible to access off a type structure derived from an unbounded generic, as this should be equivalent to a generic bounded explicitly by `mixed`.

Reviewed By: vassilmladenov

Differential Revision: D41773685

fbshipit-source-id: a6719eeb6778cb2b4175067daf43d2c0e81e3a6e