Add a new Autoload.UserlandEnabled option to make autoload_set_paths() throw
Summary: Native Facts' ultimate deliverable is the removal of `HH\autoload_set_paths()` from HHVM code, unblocking HHVM features that depend on native autoloading. Today, calling `HH\autoload_set_paths()` installs a userland autoloader on top of the native autoloader, disabling native autoloading for the rest of the request. As we roll native autoloading out everywhere, we'll want to be sure that `HH\autoload_set_paths()` is not overriding the native autoloader. One way of doing this is with a runtime option which causes us to throw when we call `HH\autoload_set_paths()`.

Reviewed By: DavidSnider

Differential Revision: D41828470

fbshipit-source-id: a66a3175318fd566cb28827ea05de584570382ff