Update serializer extension to apply type wrappers
Summary:
Related hack diff - D38654757

A field with wrapped type may be nullable. For eg :

```
wrapper
typedef Foo wrapped_struct;

Bar {
1: wrapped_struct field__;
}

```

`field__` will be of type `?MyWrapper<Foo>`.

As a result, during serialization, we will not know which wrapper to use. Unlike adapters, we cannot provide the wrapper name in the field spec because wrapper is a parameterized class.

We added a thrift only setter in the struct for such fields. We will use this setter for wrapped types during serialization.

During deserialization, we will simply check if the field is of wrapper type, we will use the wrapper methods to get the underlying value.

#forcetdhashing

Reviewed By: prasad223

Differential Revision: D42015221

fbshipit-source-id: 955485d609816f84232f1134f515ea5f2abfc2f8