--find-refs for members: reduce the number of files searched for references
Summary:
What we would do before this diff: We'd get all the descendants and everything depending on the descendants.

What we do now: We use the same fanout as for a member change impacting hierarchy, e.g. member adding or removing. That means: collecting descendants, then using hash-mingling to get the depgraph nodes for the member in the descendants and get the dependents of those.

Reviewed By: ljw1004

Differential Revision: D41691232

fbshipit-source-id: 7fb051a111768933ad4cbda00e5da9006c540d2a