Close socket in between tests in socket_sendto_longpath.php
Summary:
This test defines a socket at a random path, longer than the maximum pathlength for a particular operating system. It then tries to open the socket four times with differently-sized substrings. Right now, the test is failing with an "Address already in use" error

This occurs on each test after the first test that successfully opens the socket. I think this means the problem is that we don't close the socket after every test, and reopening the socket isn't working in our test environment.

As an alternative to, or in addition to, closing the socket, we could generate a new socket with each test, increasing their isolation level.

Reviewed By: ottoni

Differential Revision: D42053162

fbshipit-source-id: 72a202c6067fa2ac36bbc227a1e15d1b8a9c488b