shmrs supports flushables
Summary:
Currently there is support for storing shallow and folded decls in the same heap. We want to decouple the decl store from the depgraph to ease architectural burdens. So when a file changes, instead of querying the depgraph for impacted derived decls, we would simply invalidate (flush out) all folded decls from the heap and recompute them. This diff adds support for marking an entry in sharedmem as flushable.

To that end, I added an additional struct `ShmEntryData`. It's a wrapper around the original data but on top of that carries information about whether the data is a shallow decl or a folded one – a folded decl is flushable but a shallow decl is not. I'm also using one of the free bits in a `ShmemEntry` header to store information about flushability.

Hence, when we wish to store a decl, we wrap it with `ShmemEntryData::as_folded` or `ShmemEntryData::as_shallow` to signal its type. And when we wish to contruct a  `ShmemEntry` from a `ShmemEntryData`, we take its data and use its type to set the bit in the header. (See hphp/hack/src/facebook/hh_decl/decl_mmap.rs for the main logic).

Note that this diff only marks a sharedmem entry as flushable or not; there's no flushing yet – I'll implement that in a later diff.

Reviewed By: ljw1004

Differential Revision: D41812309

fbshipit-source-id: c292c3582751606501084d3e4f1d35798d08ebc4