require hh_fanout to start with dep graph
Summary:
This is a large refactor of `HhFanoutImpl` to simplify the component to always assume we start with a dep graph. `HhFanoutImpl` will now only work given a dep graph. We can make this assumption now that we have `hhdg_builder`, which handles the no-dep-graph case.

This refactor came in two parts that I had difficulty separating:
1) I removed `FullDepGraphState` completely.
2) I removed `state.rs`, which had the `State` and `LockedState` abstractions, and refactored that logic into `HhFanoutImpl` with `file_rwlock`.

For the time being, I've removed `save_state()` because now `HhFanoutImpl` takes an external dep graph instead of producing one, so I'm not sure what `HhFanoutImpl` should actually save. Maybe just the delta?

Notes for the reviewer:
- Start from `hh_fanout_lib.rs`. The implementations are largely the same as the ones that existed in `state.rs` and `full_dep_graph::FullDepGraphState`, except now they use `file_rwlock`.
- Other files shouldn't have any logical changes -- it's mostly housekeeping so things still work with the changes to `HhFanoutImpl`.

Reviewed By: ljw1004

Differential Revision: D41741460

fbshipit-source-id: af58389fe4e4ef3cf7fa2ac2387f238472a2e882