Treat private methods in SDT classes as SDT
Summary: Currently we only check *private* methods in SDT classes as though they are called statically i.e. not through `dynamic`. But we permit such methods to be called through `dynamic` or through like types. This is unsound. For now, strengthen the check so that `private` and non-`private` methods are checked similarly, requiring that they be callable dynamically.

Differential Revision: D42094999

fbshipit-source-id: 6294f4d3faccbbde097f766ad928d32e4bb084ea