Hash shm_store keys using md5
Summary: We've been using rustc_hash here, which is fast, but not particularly concerned with avoiding collisions. We're seeing some erroneous typecheck errors because of collisions in the wild: for instance, the classes `EntDimARCreatorsPrivacyPolicy` and `EntDimIGCreatorsPrivacyPolicy`. Use the first 8 bytes of the md5 (as hh_shared does) instead to avoid collisions.

Reviewed By: shayne-fletcher

Differential Revision: D42084913

fbshipit-source-id: 0eb1718ef8a4fc3d8ceb53eb81acd6d347bab523