Prefer transient DashSet building the in-memory graph
Summary:
We were maintaining an unnecessary DashSet field containing all of the unique
dependent hashes seen, when those same hashes are already in the DashMap
defining the edges.

Instead, produce this DashSet transiently, when we walk all of the edges, then
throw it away. This is both a simpler data structure and theoretically a bit faster,
because we're only inserting nodes that survived the interning to get into each
DashSet edge list.

Reviewed By: edwinsmith

Differential Revision: D42089636

fbshipit-source-id: d653ea03a34ff16109ff501bf25fd21cf845b22a