Remove misleading annotations
Summary: The `Immutable` annotation gives a misleading impression that the fields it annotates are somehow immutable in Thrift while they are not. It is always possible to change field values and the implications of doing it depend on specific APIs so this information belongs in the documentation of the relevant APIs, not in annotation.

Reviewed By: iahs

Differential Revision: D42101657

fbshipit-source-id: e6e850386a4c6c8e375139624cd8e28c176db9b3