Accesses of non-AttrInitialSatisfiesTC properties aren't effect-free
Summary:
If a property isn't AttrInitialSatisfiesTC, it means creating an
object of that class may throw. If the property's type-hint definitely
does not match the initial value, any code accessing the property must
be unreachable, because an instance will never be created. Ideally we
should mark all methods as unreachable in that situations, but that's
complicated.

This can cause monotonicity violations. If the initial value doesn't
match the type-hint, it's value before analysis will be Bottom. If we
see a set to the property with type T, it's value will then be T. If
we see an access of the property, we'll mark it as effect-free and
returning T. If a later round of analysis no longer sees the set
(because it's unreachable), we'll now think the property has type
Bottom. When we analyze the access, we'll mark it as not effect-free
(because the access is unreachable). This is a monotonicity violation
because we went from effect-free to not effect-free. Work around this
by not treating an access to a property which isn't
AttrInitialSatisfiesTC as effect-free.

Reviewed By: mdko

Differential Revision: D42112868

fbshipit-source-id: 1070b4dc313f4388d3156425bf753d2c6f72f515