Properly mark more situations as unreachable in HHBBC
Summary:
When HHBBC generates a bottom, that means the action will throw (or
fatal), and any subsequent code is unreachable. It is required to call
unreachable() in this situation to let the interpreter know
that. However, in a few situations (namely when doing various
arithmatic and bitwise operations on strange types), we were producing
a bottom but not calling unreachable(). This causes later assertions
when we try to put the bottom in an array or otherwise us it. Call
unreachable() in these circumstances.

Reviewed By: ottoni

Differential Revision: D42111732

fbshipit-source-id: ac5877ab48dde3d87ed8a2f57dc2d0c79991fb7d