Fixing memory leak when compression is enabled
Summary:
FBLite is seeing memory leak.
S309700: FBLite Infra Short term SLO violation - Proxy health

When compression is enabled, inflate method is called to unzip the payload. Unzipped ByteBuf (a new buffer created in off-heap memory) is returned and is released at the end of the processing, however, the input ByteBuf is not released.

With this diff, input byte buf is also released.

Reviewed By: robertroeser

Differential Revision: D41955619

fbshipit-source-id: 92e3b6fcd9f748f19a02e98d2e57fc317d997988