Fix HHBBC type unioning between specialized/non-specialized arrays
Summary:
In some situations we weren't unioning together specialized array
types with non-specialized array types.

For example: SVecN("a") | TSKeysetN should be
{TSKeysetN|SVecN}[UncArrKey:UncArrKey]. Instead we were producing
TSKeysetN|SVecN. In other words, we were dropping the
specialization. This isn't "incorrect", but the union is wider than it
should be. Moreover, a different order of unions can preserve the
specialization, breaking monotonicity.

Several of the operators were just discarding the specialization. Add
the logic to preserve the specialization as needed.

Reviewed By: ottoni

Differential Revision: D42113296

fbshipit-source-id: f1a58ad0d4c9b17c469c31d4684d1fd5f22e913a