Properly type SplPriorityQueue
Summary:
`SplPriorityQueue<T>` is missing a bunch of types, and properly typing its methods requires adding a second generic parameter - one for the value type, and one for the priority type.

I'm not totally sure how best to handle the WWW migration. If this diff is accepted, I plan on adding FIXMEs to the relevant callsites in WWW, then scheduling a task to clean the FIXMEs up after this diff is pushed in the next Hack release.

Reviewed By: rodmk

Differential Revision: D42079783

fbshipit-source-id: d31017c2967a58a59bf7a3f469d7a88ef6dd87dc