Support nonstandard cpp types when using cpp.indirection
Summary: Indirection types can expose a different type than the standard type (e.g. uint instead of int), and because op::decode uses outparams this causes compilation errors. Add a helper that wraps the tag when necessary. This is a consequence of cpp.indirection abusing the cpp.type annotation, and can be reverted once cpp.indirection is fully deprecated.

Reviewed By: vitaut

Differential Revision: D42156057

fbshipit-source-id: 178d43300c6691f075e82fc41cb3591daa53cf5b