Remove another misleading annotation
Summary: Similar to D42101657 (https://github.com/facebook/hhvm/commit/fb9e5ad65689efe4345e3c96511839b31a7510d5): The `Unique` annotation gives a misleading impression that we somehow enforce uniqueness of these fields in Thrift while we are not. Uniqueness guarantees are provided by specific APIs and the context in which they are unique varies. Therefore this information belongs in the documentation of the relevant APIs, not in an annotation.

Reviewed By: iahs

Differential Revision: D42148851

fbshipit-source-id: f5944a66466868579b1ee20086837f6edf583351