shmrs supports "flush"
Summary:
The last diff added ways for us to determine whether a piece of data in shmem is flushable. This diff implements the functionality to "flush". It does that by adding a 32-bit header to the front of each blob stored in shmem, which tells us
1. whether this blob is valid or not; if not, it will not be retained during the compaction phase
2. the size of the blob, so that as we iterate through shmem during compaction, we'll know at the start of each blob where the next one lies

The "flush" operation is simply to flip the first bit in the header from 1 to 0, marking the blob that follows as invalid.

Reviewed By: ljw1004

Differential Revision: D41991541

fbshipit-source-id: bdab646157cfcb0204071e709fa7641ff1986090