Produce type for default ctor in TAST
Summary: When producing a node for `new A(...)` in the TAST, the type checker currently inserts `Terr` (error type) for the type of the constructor, if `A` doesn't have an explicit constructor. This is misleading; better to emit the nullary void function type, which is the true type of the implicit constructor.

Reviewed By: Wilfred

Differential Revision: D42173177

fbshipit-source-id: c8656f6733f5cc17135f5a6e022143cfe12e33fb