Add an optional, more efficient encoding for StringData with BlobEncoder
Summary:
Unless if you're using the special UnitEmitter support, the
BlobEncoder for StringData just writes out the string verbatim. If you
have a lot of duplicate strings, this becomes very wasteful. Support
an alternate encoding which is more efficient. The first time a string
is encoded, encode the string as normal and assign the string an index
implicitly. If the same string is encoded again, simply emit the index
(which is a small integer). This is similar to how various things are
encoded in the jumpstart file.

This encoding is opt-in. It's not enabled by default because it only
works if you decode a blob precisely in order. This isn't always the
case, as BlobDecoder has support for decoding parts of the blob out of
order. Also, if you emit a lot of unique strings, the encoding is
slightly larger. So, you must explicitly enable it.

Reviewed By: edwinsmith

Differential Revision: D42161430

fbshipit-source-id: e83b873bb02aed5344b4f056fc981adc74326b68