Fix old incorrect assert in DepGraph reader
Summary:
The DepGraph's "indexer" section starts with an a 8-byte count of how
many Dep values follow. So if there are zero Deps (an empty graph) this
section needs only 8 bytes, for the count.

However, the code had a bad assert that at least 16 bytes must be present.
This works today by luck because the file format rounds up its size to a page
boundary, so we always have thousands of bytes, even for an empty graph.
But it breaks my rewrite in the next diff where I don't do that.

I'm guessing this bug is very old copypasta from the header check, which
checks to make sure there's room for two u32 values.

Reviewed By: edwinsmith

Differential Revision: D42199021

fbshipit-source-id: cf916b0adecf2f58061d4d8b88cff06ae71cdd99