Refactor and optimize .hhdg generation
Summary:
The purpose of this diff is to refactor .hhdg generation so it's possible
to plug in a new .hhdg file format. But there are some nice side effects
to cleanups made along the way:

- *.hhdg_delta -> .hhdg time dropped from 140s to 100s, a 40% speedup.
- The speed of ingesting .hhdg_delta files dropped 46s -> 34s, and produces
  a table of unique Dep hashes while doing so.
- I cut the max resident memory almost in half (25.9 GiB to 14.3 GB)
  by using 32-bit HashIndexes internally rather than full 64-bit hashes.
- `Edges::finish()` now produces a `MemDepGraph`, which is essentially
  the same thing as our current on-disk `.hhdg` file. The idea is to feed
  that into back ends that write out on-disk files.
- I ripped out the old .hhdg writer and replaced it with a trivial one --
  there's almost nothing for it to do now that`MemDepGraph` exists. Writing
  is no longer parallel so could probably be optimized, but since I'm working
  on replacing that file format and the overall perf is already better I
  didn't bother.
- Minor: I cleaned up logging to only log per 100 files ingested rather than
  for every one.

## A bit more detail on the improved inefficient `DepGraphDelta` reading:

A. The .hhdg_delta file nicely groups each dependency with any
number of dependents, but the parsing code took no advantage
of that information. Instead when it built up a
`HashMap<Dep, HashSet<Dep>>`, it inserted each (dependency, dependent)
pair one at a time, looking up the same HashSet for that dependency
again and again for every dependent in the list.

B. We then dumped this HashMap into `Edges`, which had essentially the
same data structure, a `DashMap<Dep, HashSet<Dep>>`. And *again*
we completely forgot that one dependency maps to everything in
its HashSet of dependents, and inserted (dependency, dependent)
pairs one at a time, re-taking a DashMap lock to look up the same
dependency every time.

Now, we skip that middle HashMap (in DepGraphDelta) and dump everything
straight into the final table. And we just look up the dependency's HashSet
once, and dump in all of the dependents from the original DepGraphDelta
list on disk into it in one go.

Relatedly, in a previous diff I switched the DepGraphDelta file
format to native endian so a bunch of junk could be deleted and we
point directly into mmapped files.  We should stop using big endian
everywhere because it inhibits clean mmap-based optimizations (we
already weren't using it in .hhdg files).

## Notes on .hhdg_delta reading:

Now the `.hhdg_delta` reader converts 64-bit Dep values to 32-bit HashIndex values
as it goes, and uses those for dependents sets. Those numbers are nondeterministically
produced by incrementing an atomic, so once we're done and Edges::finish() is called
we sort the Deps and compute a mapping table from old -> new 32 bit value and update
every Dep in parallel to use the new numbering.

Reviewed By: edwinsmith, nt591

Differential Revision: D42132543

fbshipit-source-id: 284e99b26eacd8f78a2368fbbad1b1ec2cf7967a