Generate class func families using extern-worker in HHBBC
Summary:
Convert the "define func families" pass over to use extern-worker. We
can do this at the same time as calculating the rest of the data in
BuildSubclassListJob. In fact, once we have the data for func families
we can actually use it for some of the other calculations (like
determining which methods are AttrNoOverride).

In extern-worker we create FuncFamily2, which is related to FuncFamily
just as ClassInfo2 and ClassInfo as related. That is, FuncFamily2 will
eventually replace FuncFamily when everything is converted over to
extern-worker. Since FuncFamily2 exists in extern-worker, we cannot
just use a pointer to represent a FuncFamily. Every FuncFamily2 has an
assigned id. Since a FuncFamily is ultimately defined by the methods
it contains, the id is just the SHA1 hash of all of its methods in
some canonical order.

Multiple extern-worker jobs can create the same FuncFamily2 (there is
no way to prevent this unless the jobs communicate with each
other). This is fine, as they will be de-duped on the local side.

Since we may create many FuncFamilies2, it can put quite a strain on
the extern-worker infrastructure. To reduce this, bundle the func
families together into groups. Each group contains multiple func
families, but they will all be stored together as one blob. This
greatly reduces the total number of blobs needed, at the cost of
writing/reading more data on average (since when you now have to load
all func families in the group even when you needed just one).

To make the BuildSubclassListJob more efficient with this added work,
the scheduling algorithm was tweaked. Whereas before we find "root"
classes with dependencies and "promoted" some dependencies in each
job, now all classes are "roots". As before, they'll pull in any
needed dependencies, but those dependencies will never be promoted, as
every class (except leafs) will be a root somewhere already. This
generates more work overall, but allows for more classes to be
processed in parallel, leading to better overall performance).

Since dependencies are not promoted anymore, there's little benefit
for the caching of Data within a job, so that has been
removed. Instead of building up the Datas recursively, each "top"
class just iterates over its subclasses and builds it. This simplifies
the algorithm and reduces a lot of overhead from copying MethRefSets
around.

The HasPrivateAncestor bit on MethTabEntry is no longer needed, as the relevant information can inferred from the state we use to build the func families now.

This diff only changes the calculation for class func
families. Name-only func families are still generated the old way and
will be addressed next.

Reviewed By: edwinsmith

Differential Revision: D42187700

fbshipit-source-id: 434b935ead13454fe35bb6dbbf712b25ab747c95