Support dynamic type in coverage checking
Summary: The `--color` and `--coverage` options in `hh_single_type_check` and `hh_client` provide a limited measurement of what code is "typed" and what code is "untyped". Until now, this has been based on `Tany`, but going forward we plan to eliminate `Tany` and move instead to a sound type system that makes use of `dynamic` and like-types to support "untyped" code. In preparation for this, we extend the notion of "untyped" used by these options to include `dynamic`, and update the tests so that (a) they actually pass the type-checker and (b) make use of `dynamic` instead of absent type hints.

Reviewed By: vsiles

Differential Revision: D42278867

fbshipit-source-id: f6c106669f3a34c491710937e656d1fd7ea1888c