Unify `.hhdg_delta` and `.bin` file formats
Summary:
- Hack `.bin` files contain pairs of (dependent, dependency) pairs
in big endian byte order.

- Hack `.hhdg_delta` files contain native endian sequences of:
dependency with the high bit set, followed by any number of
dependents with the high bit clear.

So `.hhdg_delta` files can represent `.bin` files in the same size,
because a single edge is 16 bytes either way. So just change the `.bin`
format to be the same as the `.hhdg_delta` format, so we can unify
the reader code.

In a subsequent diff, the `.bin` suffix should be retired, but I don't
know everything that would need to be updated for that.

Reviewed By: zhyty

Differential Revision: D42209774

fbshipit-source-id: 70d9fe8c81e8d329670bff5b2e211cc62da3af77