Fix self-referencing class constant bug
Summary:
There is a bug where HHVM fatals with
`Cannot declare self-referencing constant
X'` on class constants that are
not self-referencing. This is more likely to trigger on really large constants, and
happens when constant resolution is interrupted by the IntervalTimer.

This is an issue because of how we detect recursive constants. We maintain a static
 array per class in RDS that is a mapping from class constant name to TypedValue.
When we start resolving a constant, we set the value of this constant to a sentinel bit.
If we see this constant again while resolving it and see its sentinel bit is set, we know
its recursive. This breaks when we get interrupted in the middle of this process, since
we can arbitrarily clobber/read from this map in the callback.

We need to delay running callbacks until we are no longer resolving class constants.
We check surprise flags on `ReturnHook` and `Enter` of an `86cinit`, so we can
suppress callbacks when we are in this function. I think this is enough to make this safe.

Reviewed By: ricklavoie

Differential Revision: D42079245

fbshipit-source-id: f03512ef1d11f326f8173660b8c451dc867c673e