Fix find_by_* functions in the op library
Summary:
`find_by_ordinal` and `find_by_field_id` are completely broken in a nondeterministic, hard to debug way because their implementation, `for_each_ordinal_impl`, fails to initialize its result. So if nothing was found the result of the search depends on whatever garbage happens to be on the stack. Fix this by value initializing the result.

Stumbled upon this issue when trying to debug test failures in D42147642.

Reviewed By: Mizuchi

Differential Revision: D42182027

fbshipit-source-id: cc542c505890e5f8ce69821b423c529e5ea4b817