Hide any/type fields that don't compile under py3
Summary: These files currently don't build in py3 so they can't be used directly, but their reverse dependencies are used which requires py3 to be enabled for these files. To minimize the number of py3.hidden annotations required we hide individual fields so the structs are still available to other files.

Reviewed By: thedavekwon

Differential Revision: D42220820

fbshipit-source-id: 19a3cf7af0e25bb0b465de5f2b6d3870a2e7d8fb