do not register fd for multishot/long lived
Summary: Long lived requests that have registered fds can leak files (until a kernel fix comes in). For now theres not much benefit to having registered fds on multishot requests, as the file is only taken once upfront and not for each result, so just do not use it there.

Reviewed By: dmm-fb

Differential Revision: D41028626

fbshipit-source-id: 1ae9a323a6a4ca60232d7868294e16962c527974