Remove Terr and use fresh type variables instead
Summary:
We use a special type `Terr` for ill-typed expressions for which there is no obvious "default" type. `Terr` is compatible with all other types, and is a bit of an anomaly (along with `Tany`) in our type system.

Better to remove it and use instead a fresh type variable that stands for "the type that the expression would have had if it had been well-typed". Another way of looking at it is that it's as though the badly-typed expression had simply been replaced by a `throw`.

In order to preserve some of the existing behaviour of `Terr`, where we avoid error propagation in a number of places, we introduce a new reason `Rtype_variable_error` for type variables generated for this purpose. We occasionally test for this, via the helper `Typing_utils.is_type_error`.

Reviewed By: vsiles

Differential Revision: D40585926

fbshipit-source-id: db9bf0e9007ca3ccba167421ede15fe85e5a2b0e