Remove `nohash_hasher::IsEnabled` tag
Summary:
This diff has no effect beyond deleting an unused annotation that would
allow `nohash_hasher` to be used inappropriately.

`nohash_hasher` is dangerous because it requires good entropy in the high and low bits.
The high 7 bits are used for hashbrown SIMD, and bits just below that are used to
select a DashMap shard.

Our Dep types have the high bit clear, meaning only half the SIMD tags will be
used, and the low bit also isn't random.

It looks like `nohash_hasher` was originally added for `Checksum`, where it's legit,
but then folded into various `Dep`-like types as well via a macro.

To avoid the poor hashing footgun, and an apparently unused `nohash_hasher` dependency,
just delete this marker.

If you want faster hashing for Dep-like types, the right approach IMHO is in D42096971.
It uses a bit rotate to move the junky high and low bit of Deps out of the way
but no other hashing cost. It's currently unpublished because I haven't found a case
where it clearly makes things faster.

Reviewed By: rjbailey

Differential Revision: D42221845

fbshipit-source-id: 936910af9eeac65bb723792f0faf6ff890b04c6f