Add support for `__Memoize(#Uncategorized)`
Summary:
This adds a new MemoizeOption that can be passed. It has the same memoize behavior as not passing any arguments. The only difference is that (a) it shows up in reflection and (b) you cannot use it on functions as capable as or less capable than `[leak_safe, globals]`.

The purpose of this is to eventually *require* that a memoize kind is passed for functions that are not as capable as or less capable than `[leak_safe, globlas]`. Before we can require it, we must allow it and codemod existing code.

Reviewed By: oulgen

Differential Revision: D42271572

fbshipit-source-id: 66c37a3e4f9c7e3217f3b83b4328e776c50e2c79