Fix shadowed variable in folly/experimental/coro/test/PromiseTest.cpp
Summary:
LLVM-15 has a warning `-Wshadow` which we treat as an error because it's so often diagnostic of a code issue. We fix the this error here.

Shadowed variables are variables in an inner scope with the same name as another variable in an outer scope. Having the same name for both variables might be semantically correct, but it can make the code confusing to read! It can also hide subtle bugs.

This diff fixes such an issue by renaming the variable.

(1 file changed.)

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: Orvid

Differential Revision: D42348650

fbshipit-source-id: aa9add796e0e168e2d3b72a39ee412fda088dd10