Show SDT attribute for hover on functions and methods
Summary:
When hovering over a call to an SDT function or method in an SDT class, the experience is subpar, as the SDT attribute has been converted into a use of `supportdyn<_>` that just wraps the function type.

So simply strip off the `supportdyn` and then print it as an SDT attribute.

Reviewed By: vassilmladenov

Differential Revision: D42348270

fbshipit-source-id: 1779134566c01e82060f3a11c654251fdabeea32