Don't always throw away supportdyn in array index
Summary: For `e1[e2]` if `e1 : supportdyn<t>`, the array index check just assumed that `e1 : t`, but this isn't flexible enough when we are in the mode where dynamic-aware subtyping is enabled. In that case, we want to use the fact that `e1 : dynamic`. A similar thing happens with `e1[] = e2` and `e1[e2] = e3`.

Reviewed By: andrewjkennedy

Differential Revision: D42344978

fbshipit-source-id: ec1de4789fbcb233e710766ffd6a7382978494f9