Store closure_class_name on Efun
Summary:
The name on `fun_` is used a bunch of different ways: it stores the name for global functions, but for anonymous functions it has the value `";anonymous"` or the name of the generated class after closure conversion.

This makes it hard to provide relevant hover information on anonymous functions.

Instead, add an explicit `closure_class_name` field to `Efun`, and use that for closure conversion.

Reviewed By: hgoldstein

Differential Revision: D42285655

fbshipit-source-id: dd05a05dc21ef20f02fbd71506e25fb9378d0afe