Remove eager initialization of OmniRegistry
Summary:
This has been added as a fix for conformance test, however this makes it impossible to import OmniRegistry in the Thrift Adapters as it would lead to recursively loading other thrift files prior to actually loading the adapter.

Instead of doing eager initialization everywhere, moving this to the conformance test server to load it before we start the server.

Reviewed By: junzh0u

Differential Revision: D42300053

fbshipit-source-id: acd85e62298f522c414e1d37c4e1123a5bf3d484