Support nonstandard integral types when using cpp.indirection
Summary:
Indirection types can expose a different type than the standard type (e.g. uint instead of int), and because op::decode uses outparams this causes compilation errors. Add a helper that wraps the tag when necessary. This is a consequence of cpp.indirection abusing the cpp.type annotation, and can be reverted once cpp.indirection is fully deprecated.
Resubmit after in-place clear regression was fixed.

Reviewed By: Mizuchi

Differential Revision: D42371878

fbshipit-source-id: 030fc889532d087481cf53e856a1a5684ebecad3