Introduce AnyRef and AnyConstRef
Summary:
This adds missing support for `AnyRef` and `AnyConstRef` (as per TODO) that can hold references to Thrift values. The new types are interoperable with `AnyValue`, i.e. they can be initialized either directly from Thrift object or from type erased AnyValue.

The interface is minimal and is compatible with `type::Ref` and `type::ConstRef`, in case we eventually want to replace it with those (right now those types are not compatible with AnyValue).

Reviewed By: Mizuchi

Differential Revision: D42061978

fbshipit-source-id: 2586162f5c58dd9a7bfbfe3e31ecde029b7a0ca5