Calculate name-only func families using extern-worker
Summary:
In addition to calculating the normal func families, the
BuildSubclassListJob now also calculates "name-only" func families
(func families for all methods with a particular name). Since the
name-only func families each job produces are just a subset of the
total, we need an additional job afterwards which aggregate entries
together into one.

As an optimization, we only consider methods which have the "first
name" bit set. This indicates that no parent of the class has a method
with the same name. The func family for such a method will
automatically contain any data from child classes, so we only need to
consider those.

Reviewed By: kavoor

Differential Revision: D42347755

fbshipit-source-id: 8df3f0d73c94a9891f7049a96f03a45f7da1674c