sort type xrefs
Summary:
There are two XRefs predicate in the Hack schema. FileXRefs and Typeinfo. The former for xrefs relative to a file, the latter for xrefs relateive to a pretty-printed type.

In the indexer, there are two corresponding json fact creation function. The first one sorts its elements, but not the second one. It's not incorrect since there's no specified order for these arrays, but this made diffing of DBs noisy.

Fix here is to factorize these two functions  (and now both of them sort the xref array). Their inputs are slightly different so it requires a bit of plumbing.

Reviewed By: simonmar

Differential Revision: D42402664

fbshipit-source-id: 411a5ae415470e94d927d2b9a3a57ab6850c8a78