Assignment to inout parameters during dynamic check
Summary:
Because of the way that we adjust the parameter hint when checking in dynamic mode, assignments to `inout` parameters were rejected. Specifically, Hack was rejecting an assignment of a like type `~C` to an `inout` parameter of type `C`, when in dynamic mode. It turns out that it's not necessary to adjust the parameter hint after all.

I've used the opportunity to factor out some common code in parameter checking.

Differential Revision: D42412619

fbshipit-source-id: 1de52a160b1c4678e44a2e4a209bb53a587514ef